Add files via upload
A working version of the pareto optimizatio code. The class and methods are defined in population.py. main.py starts the iterations. If the iterations have already been started and one wants to continue them is possible to do so by using continue_iterations.py. runAstra_pareto.py is the wrapper to call Astra and the necessary files for that are stored in Astra_files.